Altered LW_kernel
CTP dimension is flipped (ascending order, 90 to 900 hPa) with respect to the dimensions of the MODIS histograms, as output by most GCMs. This kernel has flipped both CTP and CTP_bin_edges to align with GCM output. Otherwise, CRK decomposition will lead to false results. This also aligns the LW kernels with the SW kernels, which have the CTP dimension in descending order.